fix: device imports using strategies
